% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alstructure_summary}
\alias{alstructure_summary}
\title{Function for writing summary of \code{alstructure}}
\usage{
alstructure_summary(run_time, out_folder)
}
\arguments{
\item{run_time}{the amount of time ALStructure tood to run}

\item{out_folder}{folder to write summary function}

\item{method}{one of \code{cALS} or \code{uALS}}
}
\value{
None
}
\description{
Writes a file "summary.txt" to the file specified by the argument
"out_folder" that includes details about the ALStructure algorithm execution
including method used and time elapsed during exectution.
}
\keyword{internal}
