% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{best_perm}
\alias{best_perm}
\title{Finds the best permutation of the rows of \eqn{\boldsymbol{\hat{Q}}}{Q_hat}.}
\usage{
best_perm(Q, Q_hat)
}
\arguments{
\item{Q}{the fixed matrix to which \eqn{\boldsymbol{\hat{Q}}}{Q_hat} is compared}

\item{Q_hat}{the matrix for which we try to find the optimal row permutation}
}
\value{
a \eqn{d \times d}{d x d} permutation matrix
}
\description{
Returns a  \eqn{(d \times d)}{d x d} permutation matrix \eqn{\boldsymbol{A}}{A} that
minimizes the Frobenius distance between \eqn{\boldsymbol{Q}}{Q} and
\eqn{\boldsymbol{A}\boldsymbol{\hat{Q}}}{Q_hat}:
\deqn{||\boldsymbol{Q} -
\boldsymbol{A\hat{Q}}||_F = \sqrt{\sum_{(i,j)} (q_{ij} - \boldsymbol{a}_{i \cdot} \boldsymbol{\hat{q}}_{\cdot j})^2 }}{|| Q - Q_hat ||_F}
This is useful when trying to make meaningful comparisons between the output
of two methods for estimating global ancestry, or for comparison to a
ground truth (when evaluating performance on simulated data). Since for any permutation matrix
 \eqn{\boldsymbol{A}}{A} we have the identity
\eqn{\boldsymbol{A}^T\boldsymbol{A} = \boldsymbol{I}}{A^T A = I}, the permuted matrices
\eqn{\boldsymbol{\tilde{P}} = \boldsymbol{P}\boldsymbol{A}^T}{PA^T} and
\eqn{\tilde{\boldsymbol{Q}} = \boldsymbol{A}\boldsymbol{Q}}{AQ} provide a
model with likelihood equal to that specified
by \eqn{\boldsymbol{P}}{P} and \eqn{\boldsymbol{Q}}{Q} under the admixture model.
}
\keyword{internal}
