% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{binomial_likelihood}
\alias{binomial_likelihood}
\title{Compute the binomial likelihood of the SNP matrix given individual-specific
allele frequencies.}
\usage{
binomial_likelihood(X, F)
}
\arguments{
\item{X}{The \eqn{m \times n}{m x n} SNP data matrix}

\item{F}{The \eqn{m \times n}{m x n} matrix of binomial parameters.}
}
\value{
The likelihood and log likelihood of the data given the parameters
}
\description{
Compute the binomial likelihood of the data \eqn{\boldsymbol{X}}{X} given the
parameters \eqn{\boldsymbol{F}}{F}. The likelihood
is the probability of the data given the parameters, which under the admixture
model is:
\deqn{\prod_{ij} \binom{2}{x_{ij}} f_{ij}^{x_{ij}} (1 - f_{ij})^{2 - x_{ij}}}
}
\keyword{internal}
