% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{BN}
\alias{BN}
\title{Simulates \eqn{\boldsymbol{P}}{P} from the BN model}
\usage{
BN(BN_params, d)
}
\arguments{
\item{BN_params}{a \eqn{m \times 2}{m x 2} matrix of parameters. The first column
contains \eqn{\textrm{F}_{\textrm{ST}}}{F_ST} for each SNP, while the second
 column contains the allele frequency (\eqn{F}{F} and \eqn{p}{p}, repectively).}

\item{d}{the dimension of the latent space.}
}
\value{
an \eqn{m \times d}{m x d} matrix \eqn{\boldsymbol{P}}{P}.
}
\description{
Simulates a matrix \eqn{\boldsymbol{P}}{P} from the Balding-Nichols (BN) model.
The Balding-Nichols distribution takes two parameters \eqn{(F, p)}{(F, p)}, such that
\deqn{\textrm{Balding-Nichols}(F, p) = \textrm{Beta}\left(\frac{1 - F}{F}p, \frac{1 - F}{F}(1 - p) \right)}{Balding-Nichols(F, p) = Beta(p(1 - F) / F, (1 - p)(1 - F) / F)}
The function requires an \eqn{m \times 2}{m x 2} matrix of parameters, with each row
a pair of parameters \eqn{(F_i, p_i)}{(F_i, p_i)} for each SNP.
}
\keyword{internal}
