% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{cals}
\alias{cals}
\title{A provable algorithm for factoring \eqn{\boldsymbol{\hat{F}}}{F_hat}.}
\usage{
cals(F_hat, d, tol = 0.00001, max_iters = 1000, P_init = NULL,
  Q_init = NULL, P_samples = NULL, Q_samples = NULL)
}
\arguments{
\item{F_hat}{The estimate of the matrix \eqn{\boldsymbol{F}}{F} to be factored}

\item{d}{The dimension of the latent space. This can be estimated by the
function \code{d_estimate}.}

\item{tol}{The convergence criterion. If \eqn{RMSE(\boldsymbol{\hat{Q}}_t - \boldsymbol{\hat{Q}}_{t + 1})
< tol}{RMSE(Q_t - Q_(t + 1)) < tol}, then the algorithm halts}

\item{max_iters}{The maximum number of iterations (repetitions of steps (6)
and (7) in Algorithm 1) to be executed}

\item{P_init}{Optional initialization of P. If blank, P is initialized
with each entry an independent draw from Unif(0,1)}

\item{Q_init}{Optional initialization of Q. If blank, Q is initialized as an
all zeros matrix.}

\item{P_samples}{Optional number of subsampled of rows of P.
When left unspecified, \code{P_samples} is set to \eqn{m}{m}. When specified, the
\code{cALS} algorithm performs the optimization over a random subset of size
\code{P_samples} during each iteration. This can substantially speed up computation.
Only available for \code{cALS} method.}

\item{Q_samples}{Optional number of subsampled of columns of Q.
When left unspecified, \code{Q_samples} is set to \eqn{n}{n}. When specified, the
\code{cALS} algorithm performs the optimization over a random subset of size
\code{Q_samples} during each iteration. This can substantially speed up computation.
Only available for \code{cALS} method.}
}
\value{
A list with the following elements:
   \describe{
   \item{P_hat}{: The estimated \eqn{\boldsymbol{P}}{P} matrix. Each
   column of \eqn{\boldsymbol{P}}{P} may be interpreted as a vector of
   allele frequencies for a specific ancestral population.}
   \item{Q_hat}{: The estimated \eqn{\boldsymbol{Q}}{Q} matrix. Each
   column of \eqn{\boldsymbol{Q}}{Q} may be interpreted as the admixture
   proportions of a specific individual.}
   }

 @references
 Cabreros, I., and J. D. Storey. 2017. “A Nonparametric Estimator of Population Structure Unifying Admixture Models and Principal Components Analysis.” BioRxiv. Cold Spring Harbor Laboratory. doi:10.1101/240812.
}
\description{
An algorithm for finding approximate factors \eqn{\boldsymbol{\hat{P}}}{P_hat} and \eqn{\boldsymbol{\hat{Q}}}{Q_hat} of the matrix \eqn{\boldsymbol{\hat{F}}}{F_hat} that obey constraints of the admixture model:
\enumerate{
 \item \eqn{p_{ij} \in [0,1]  ~ \forall (i,j)}{p_ij in [0, 1] for all (i, j)}
 \item \eqn{q_{ij} \ge 0  ~ \forall (i,j)}{q_ij >= 0 for all (i, j)} and \eqn{\sum_i q_{ij} = 1 ~ \forall j}{q_1j + q_2j + ... + q_dj = 1 for all j}
}
This algorithm is described in Algorithm 1 in (Cabreros and Storey 2017) and
has the provable guarantee of asymptotically converging to a stationary
point of the objective function.
}
\keyword{internal}
