% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{compare_Q}
\alias{compare_Q}
\title{Generate a scatterplot of rows of \eqn{\boldsymbol{Q}}{Q} and \eqn{\boldsymbol{\hat{Q}}}{Q_hat}.}
\usage{
compare_Q(Q, Q_hat, output_file = NaN, dims = c(1, 2), res = 100,
  rand_held_out = 0)
}
\arguments{
\item{Q}{The "true" \eqn{d \times n}{d x n} underlying admixture matrix}

\item{Q_hat}{The estimated \eqn{k \times n}{d x n} admixture matrix to compare to Q}

\item{output_file}{The file to which the generated plot will be written}

\item{dims}{A vector of length 2 that specifies which 2 rows will be plotted}

\item{res}{The resolution of the output figure}

\item{rand_held_out}{The number of randomly selected points which will be
colored differently from the others. This feather is to allow for a
clear visual demonstration of how well individual points concur
between \eqn{\boldsymbol{Q}}{Q} and \eqn{\boldsymbol{\hat{Q}}}{Q_hat}.}
}
\value{
A \code{ggplot} object \code{p_all} containing the scatterplot of both the estimate
        and ground truth side by side.
}
\description{
Generate a scatterplot of rows of \eqn{\boldsymbol{Q}}{Q} and \eqn{\boldsymbol{\hat{Q}}}{Q_hat}
for a visual comparison.
This can be used to compare the structure of \eqn{\boldsymbol{\hat{Q}}}{Q_hat} fitted using
\code{ALStructure} to the truth (in the case of simulated data), or it can be
used to compare the outputs of two fitting algorithms. The plots are written
into the file specified by \code{output_file}.
}
\keyword{internal}
