% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{D_binomial}
\alias{D_binomial}
\title{Estimates the \eqn{n \times n}{n x n} matrix \eqn{\boldsymbol{D}}{D} for Latent Subspace
Estimation.}
\usage{
D_binomial(X)
}
\arguments{
\item{X}{the \eqn{m \times n}{m x n} SNP data matrix}
}
\value{
the \eqn{n \times n}{n x n} matrix \eqn{\boldsymbol{\hat{D}}}{D_hat}
}
\description{
Estimates the \eqn{n \times n}{n x n} matrix \eqn{\boldsymbol{D}}{D} using the estimator
from Chen and Storey (2015) (Table 1). \eqn{\boldsymbol{D}}{D} is used to adjust
for heteroskedasticity when estimating the latent subspace. This function is
used by \code{F_estimate()}.
}
\references{
CHEN, X., and J. D. STOREY, 2015 Consistent Estimation of Low-Dimensional
Latent Structure in High-Dimensional Data. arXiv:1510.03497v1.
}
\keyword{internal}
