% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_d}
\alias{estimate_d}
\title{Estimate the latent space dimension}
\usage{
estimate_d(X)
}
\arguments{
\item{X}{the \eqn{m \times n}{m x n} SNP data matrix}
}
\value{
an estimate \eqn{\hat{d}}{d_hat} of the dimension of the latent space
        dimension.
}
\description{
Estimates the dimension of the rowspace of \eqn{\boldsymbol{Q}}{Q}
 (equivalently, the rank
of the matrix \eqn{\boldsymbol{F}}{F}). This estimate \eqn{\hat{d}}{d_hat} is based on the estimator from (Leek 2011), page 6.
}
\references{
Leek, J. T. 2011. “Asymptotic conditional singular value decomposition for high-dimensional genomic data.” Biometrics 67 (4): 344–52.
}
