% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_F}
\alias{estimate_F}
\title{Estimates the individual-specific allele frequency matrix \eqn{\boldsymbol{F}}{F}}
\usage{
estimate_F(X, d, svd_method = "base")
}
\arguments{
\item{X}{The \eqn{m \times n}{m x n} SNP data matrix}

\item{d}{The rank of \eqn{\boldsymbol{F}}{F}. This can be estimated using the function
\code{d_estimate()}.}

\item{svd_method}{One of "base" or "truncated_svd." If "base" is chosen, the
base \code{svd()} function is used. If "truncated_svd" is chosen, the truncated svd algorithm \code{propack.svd()} from the \code{svd} package is used.}
}
\value{
A list with the following elements:
   \describe{
   \item{F_hat}{: The \eqn{m \times n}{m x n} matrix \eqn{\boldsymbol{\hat{F}}}{F_hat}.}
   \item{rowspace}{: a list with the following elements:
   \describe{
       \item{vectors}{: The top \eqn{d}{d} eigenvectors of the matrix
       \eqn{\boldsymbol{G}}{G} sorted by decreasing eigenvalue. These vectors
   approximate the subspace spanned by the rows of \eqn{\boldsymbol{Q}}{Q}.}
       \item{values}{: The top \eqn{d}{d} eigenvalues of the matrix
       \eqn{\boldsymbol{G}}{G} sorted by decreasing eigenvalue.}}}
   }
}
\description{
Estimates the \eqn{m \times n}{m x n} individual-specific allele frequency matrix
\eqn{\boldsymbol{F}}{F} using the method of Latent Subspace Estimation
 (X. Chen and Storey 2015) as described in (Cabreros and Storey 2017)
in Section 2.3.
}
\references{
Cabreros, I., and J. D. Storey. 2017. “A Nonparametric Estimator of Population Structure Unifying Admixture Models and Principal Components Analysis.” BioRxiv. Cold Spring Harbor Laboratory. doi:10.1101/240812.

Chen, X., and J. D. Storey. 2015. “Consistent Estimation of Low-Dimensional Latent Structure in High-Dimensional Data.” ArXiv E-Prints, October.
}
