% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factor_F}
\alias{factor_F}
\title{A fast algorithm for factoring \eqn{\boldsymbol{\hat{F}}}{F_hat}.}
\usage{
factor_F(F_hat, d, tol = 0.00001, max_iters = 1000)
}
\arguments{
\item{F_hat}{The estimate of the matrix \eqn{\boldsymbol{F}}{F} to be factored}

\item{d}{The dimension of the latent space. This can be estimated by the
function \code{d_estimate}.}

\item{tol}{The convergence criterion. If \eqn{RMSE(\boldsymbol{\hat{Q}}_t - \boldsymbol{\hat{Q}}_{t + 1})
< tol}{RMSE(Q_t - Q_(t + 1)) < tol}, then the algorithm halts}

\item{max_iters}{The maximum number of iterations (repetitions of steps (6)
and (7) in Algorithm 1) to be executed}
}
\value{
A list with the following elements:
   \describe{
   \item{P_hat}{: The estimated \eqn{\boldsymbol{P}}{P} matrix. Each
   column of \eqn{\boldsymbol{P}}{P} may be interpreted as a vector of
   allele frequencies for a specific ancestral population.}
   \item{Q_hat}{: The estimated \eqn{\boldsymbol{Q}}{Q} matrix. Each
   column of \eqn{\boldsymbol{Q}}{Q} may be interpreted as the admixture
   proportions of a specific individual.}
   }

 @references
 Cabreros, I., and J. D. Storey. 2017. “A Nonparametric Estimator of Population Structure Unifying Admixture Models and Principal Components Analysis.” BioRxiv. Cold Spring Harbor Laboratory. doi:10.1101/240812.
}
\description{
An algorithm for finding approximate factors \eqn{\boldsymbol{\hat{P}}}{P_hat} and \eqn{\boldsymbol{\hat{Q}}}{Q_hat} of the matrix \eqn{\boldsymbol{\hat{F}}}{F_hat} that obey constraints of the admixture model:
\enumerate{
 \item \eqn{p_{ij} \in [0,1]  ~ \forall (i,j)}{p_ij in [0, 1] for all (i, j)}
 \item \eqn{q_{ij} \ge 0  ~ \forall (i,j)}{q_ij >= 0 for all (i, j)} and \eqn{\sum_i q_{ij} = 1 ~ \forall j}{q_1j + q_2j + ... + q_dj = 1 for all j}
}
This algorithm is described in Algorithm 2 in (Cabreros and Storey 2017). While
it lacks the theoretical guarantees of the \code{cALS} algorithm, it is much faster.
}
