% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute an SNP matrix with NAs}
\usage{
impute_mean(X)
}
\arguments{
\item{an}{\eqn{m \times n}{m x n} SNP matrix with NAs}
}
\value{
an \eqn{m \times n}{m x n} SNP matrix with mean-imputed values.
}
\description{
Impute an SNP matrix with NAs by replacing each NA with the row mean
(excluding NAs) from each row
}
\keyword{internal}
