% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{lse}
\alias{lse}
\title{Estimates the latent subspace}
\usage{
lse(X, d, svd_method = "base")
}
\arguments{
\item{X}{The \eqn{m \times n}{m x n} SNP data matrix}

\item{d}{The rank of \eqn{\boldsymbol{F}}{F}. This can be estimated using the
function \code{d_estimate()}. When \eqn{d = n}{d = n}, all eigenvectors
of \eqn{\boldsymbol{G}}{G} are returned.}

\item{svd_method}{One of "base" or "truncated_svd." If "base" is chosen, the
base \code{svd()} function is used. If "truncated_svd" is chosen, the truncated svd algorithm \code{propack.svd()} from the \code{svd} package is used.}
}
\value{
A list with the following elements:
   \describe{
   \item{vectors}{: The top \eqn{d}{d} eigenvectors of the matrix
   \eqn{\boldsymbol{G}}{G} sorted by decreasing eigenvalue. These vectors
   approximate the subspace spanned by the rows of \eqn{\boldsymbol{Q}}{Q}}.
   \item{values}{: The top \eqn{d}{d} eigenvalues of the matrix
   \eqn{\boldsymbol{G}}{G} sorted by decreasing eigenvalue.}
   }
}
\description{
Estimates the rowspace of Q using the method of latent subspace estimation.
The function returns the top d eigenvalues and vectors of the matrix
\deqn{\boldsymbol{G} = \frac{1}{m} \boldsymbol{X}^T \boldsymbol{X} - \boldsymbol{D}}{G = 1/m X^T X - D}
where the matrix D is a diagonal matrix with each diagonal entry
\eqn{d_{ii}}{d_ii} an estimate of the average of the variances of the
random variables in the \eqn{i}{i} column of \eqn{\boldsymbol{X}}{X}. As is
proven in (X. Chen and Storey 2015), the span of the top \eqn{d}{d} eigenvectors
of \eqn{\boldsymbol{G}}{G} span the same space as the rows of
\eqn{\boldsymbol{Q}}{Q}. The eigenvectors are returned in order of decreasing
eigenvalue.
}
\references{
Chen, X., and J. D. Storey. 2015. “Consistent Estimation of Low-Dimensional Latent Structure in High-Dimensional Data.” ArXiv E-Prints, October.
}
