% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{make_bed}
\alias{make_bed}
\title{Creates \code{.bed} files from matrix}
\usage{
make_bed(X, out_file, B = min(ceiling(dim(X)[2]/10)))
}
\arguments{
\item{X}{The \eqn{m \times n}{m x n} SNP matrix of 0, 1, and 2.}

\item{out_file}{The output file. No extension should be added. Three files,
\code{out_file.bed}, \code{out_file.fam}, and \code{out_file.bim} will be created.}

\item{B}{The number of blocks to to write data in. Default is set to
\code{ceil(m / 10)}.}
}
\value{
Three files written to \code{out_file}:
         \describe{
         \item{\code{out_file.bed}}{SNP matrix in \code{.bed} format}
         \item{\code{out_file.fam}}{artificial file}
         \item{\code{out_file.bim}}{artificial file}}
}
\description{
Creates \code{.bed}, \code{.fam}, and \code{.bim} files from an SNP matrix for the purposes of comparing \code{ALStructure} to other methods. The \code{.bim} and \code{.fam} files are artificial, however we include them because other methods may require them as input.
}
\keyword{internal}
