% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{mean_ABS}
\alias{mean_ABS}
\title{Compute the mean absolute value difference between matrices \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B}}
\usage{
mean_ABS(A, B)
}
\arguments{
\item{A}{An \eqn{m \times n}{m x n} matrix}

\item{B}{An \eqn{m \times n}{m x n} matrix}
}
\value{
The mean absolute value difference between \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B}.
}
\description{
Compute the mean absolute difference value between matrices \eqn{\boldsymbol{A}}{A}
 and \eqn{\boldsymbol{B}}{B}. If \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B}
 are \eqn{m \times n}{m x n} matrices, then the mean absolute value difference
is:
\deqn{\frac{1}{mn} \sum_{ij} |a_{ij} - b_{ij}|}{1 / mn sum(|a_ij - b_ij|)}
This function is useful for comparing the performance of \code{ALStructure}
fits of global ancestry to either alternative methods or the ground truth
(in the case of simulations).
}
\keyword{internal}
