% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{the ggplot2 plot objects to be plotted}

\item{cols}{the number of columns into which the plots will be arranged}
}
\value{
a ggplot2 object with supplied plots arranged in a grid
}
\description{
Plot a colection of ggplot2 plot objects into a simple grid layout with a
specified number of columns
}
\keyword{internal}
