% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{order_pops}
\alias{order_pops}
\title{Orders the \eqn{d}{d} populations}
\usage{
order_pops(P, Q, method = "ave_admixture", Q_space = NULL)
}
\arguments{
\item{P}{The \eqn{m \times d}{m x d} loadings matrix with columns to be
ordered}

\item{Q}{The \eqn{d \times n}{d x n} admixture matrix with rows to be ordered}

\item{method}{One of "ave_admixture" or "var_explained." If "ave_admixture," the \eqn{d}{d} populations are ordered by decreasing average admixture accross samples (i.e. \eqn{1 / n \sum_{j} q_{ij}}{1 / n (q_i1 + q_i2 + ... + q_in)}). If "var_explained", the \eqn{d}{d} populations are ordered be decreasing variation explained. Specifically, we compute a modified version of the \eqn{\mbox{eigen-}R^2}{eigen-R^2} statistic from (L. S. Chen and Storey 2008). The statistic is modified in the following ways: 1) we treat
rows of \eqn{\boldsymbol{Q}}{Q} as the response variables 2) we regress
each row of Q on the eigenvectors of \eqn{\boldsymbol{G}}{G} rather than the
eigenvectors of the data matrix itself 3) we take the weighted average only
over the top \eqn{d}{d} eigenvectors.}

\item{Q_space}{Only required for "var_explained" methodThe list containing
the top \eqn{d}{d} eigenvectors and their
corresponding eigenvalues of the \eqn{\boldsymbol{G}}. These may be obtained
through the function \code{lse}.}
}
\value{
A list with the following elements:
   \describe{
   \item{P_ordered}{: The permuted \eqn{\boldsymbol{P}}{P} }
   \item{Q_ordered}{: The permuted \eqn{\boldsymbol{Q}}{Q} }
   \item{perm_mat}{: The permutation matrix \eqn{\boldsymbol{A}}{A} such that
   \eqn{\boldsymbol{PA}^T = \boldsymbol{P}_{\mbox{ord}}}{PA^T = P_ord} and \eqn{\boldsymbol{AQ} = \boldsymbol{Q}_{\mbox{ord}}}{AQ = Q_ord}.}
   }
}
\description{
Orders the \eqn{d}{d} populations according to one of two methods: "ave_admixture" or "var_explained." Function returns matrices \eqn{\boldsymbol{P}}{P} and \eqn{\boldsymbol{Q}}{Q} with permuted columns and rows according to the determined ordering of populations.
}
\references{
Chen, L. S., and J. D. Storey. 2008. “Eigen-R2 for dissecting variation in high-dimensional studies.” Bioinformatics 24 (19): 2260–2.
}
