% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{projsplx}
\alias{projsplx}
\title{Project a vector onto the simplex}
\usage{
projsplx(y)
}
\arguments{
\item{y}{a \eqn{n}{n} dimensional vector}
}
\value{
a \eqn{n}{n} dimensional vector which is the projection of \eqn{y}{y} onto
        the \eqn{n}{n} dimensional simplex
}
\description{
Algorithm from (Y. Chen and Ye 2011) to project an \eqn{n}{n} dimensional vector onto the \eqn{n}{n} dimensional
simplex (i.e. the set of points in \eqn{x \in \mathcal{R}^n}{x in R^n} such that
\eqn{\sum_i x_i = 1}{x_1 + x_2 + ... + x_n= 1} and \eqn{x_i > 0}{x_i > 0} for all \eqn{i}{i} ). This is used to enforce the admixture constraints in
the \code{uALS} algorithm.
}
\references{
Chen, Y., and X. Ye. 2011. “Projection Onto A Simplex.” ArXiv E-Prints, January.
}
\keyword{internal}
