% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{RMSE}
\alias{RMSE}
\title{Compute the root mean squared error (RMSE) between matrices \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B}.}
\usage{
RMSE(A, B)
}
\arguments{
\item{A}{An \eqn{m \times n}{m x n} matrix}

\item{B}{An \eqn{m \times n}{m x n} matrix}
}
\value{
The RMSE between \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B}.
}
\description{
Compute the root mean squared error (RMSE) between matrices \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{A}. If \eqn{\boldsymbol{A}}{A} and \eqn{\boldsymbol{B}}{B} are \eqn{m \times n}{m x n} matrices, then the RMSE is:
\deqn{\sqrt{\frac{1}{mn} \sum_{ij} |a_{ij} - b_{ij}|^2}}{sqrt(1 / mn sum((a_ij - b_ij)^2))}
This function is useful for comparing the performance of \code{ALStructure}
fits of global ancestry to either alternative methods or the ground truth
(in the case of simulations). This is also the function used in the convergence
criterion of \code{cALS} and \code{uALS}.
}
\keyword{internal}
