% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_admixture}
\alias{simulate_admixture}
\title{Simulates data from the PSD model}
\usage{
simulate_admixture(m, n, d, alpha, BN_params = NA, seed = NA)
}
\arguments{
\item{m}{number of SNPs}

\item{n}{number of individuals}

\item{d}{number of groups}

\item{alpha}{dirichlet parameter; \code{length(alpha) = d}}

\item{BN_params}{a \eqn{m \times 2}{m x 2} matrix of parameters. The first column
contains \eqn{\textrm{F}_{\textrm{ST}}}{F_ST} for each SNP, while the second
 column contains the allele frequency.}
}
\value{
a list with the following elements:
   \describe{
   \item{P}{: the \eqn{m \times d}{m x d} matrix of loadings}
   \item{Q}{: the \eqn{d \times n}{d x n} matrix of latent admixture components}
   \item{F}{: the \eqn{m \times n}{m x n} matrix \eqn{PQ}{PQ}}
   \item{X}{: a random draw such that \eqn{x_{ij} \sim \mbox{Bernoulli}(f_{ij}, 2)}{x_ij ~ Bernoulli(f_ij, 2)}}.
   }
}
\description{
Creates a data frame that contains the parameters of the admixture model
\eqn{(\boldsymbol{F},\boldsymbol{P},\boldsymbol{Q})}{(F, P, Q)} as well as a
  single draw \eqn{\boldsymbol{X}}{X} such that \eqn{x_{ij} \sim \mbox{Bernoulli}(f_{ij})}{x_ij ~ Bernoulli(f_ij)}.
  The \eqn{\boldsymbol{Q}}{Q} matrix is drawn from the Dirichlet distribution with parameter
  \eqn{\alpha}{alpha}, and the \eqn{\boldsymbol{P}}{P} matrix is simulated from the Balding-Nichols model.
  The parameter \eqn{\alpha}{alpha} is supplied by the user. The \eqn{m \times 2}{m x 2} matrix
  of Balding-Nichols parameters is optional. If \code{BN_params} is not
  supplied, the parameters are derived from a random sample of estimated
  Balding-Nichols parameters from the Human Genomes Diversity Project (HGDP)
  dataset. The Balding-Nichols parameter estimates are provided by (Gopalan et al. 2016), and included in this package in the object \code{hgdpBN}.
}
\references{
Gopalan, P., W. Hao, D. M. Blei, and J. D. Storey. 2016. “Scaling probabilistic models of genetic variation to millions of humans.” Nature Genetics 48 (12): 1587–90.
}
